# src/main/resources/static/api-docs.yml
...
openapi: 3.0.3
info:
  title: "demo-api"
  version: 1.0.0
  description: >-
    demo api.
  license:
    name: Apache-2.0
    url: 'https://www.apache.org/licenses/LICENSE-2.0.html'
paths:
...
  /refresh:
    post:
      tags:
        - authentication
      description: Refresh token
      operationId: refresh
      parameters:
        - in: header
          name: X-Forwarded-For
          schema:
            type: string
          required: false
        - in: header
          name: X-Refresh-Token
          schema:
            type: string
          required: true
      responses:
        200:
          description: Token refreshed
          content:
            application/vnd.kaiqkt_auth_api_auth_v1+json:
              schema:
                $ref: '#/components/schemas/AuthenticationResponseV1'
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
        401:
          description: Session not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
...
  /session/revoke/current:
    patch:
      tags:
        - session
      description: Revoke current session
      operationId: revokeCurrent
      security:
        - JwtToken: [ ]
      responses:
        200:
          description: Session revoked
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /sessions/revoke:
    patch:
      tags:
        - session
      description: Revoke all user sessions
      operationId: revokeAll
      security:
        - JwtToken: [ ]
      responses:
        200:
          description: Sessions revoked
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /sessions/revoke/all:
    patch:
      tags:
        - session
      description: Revoke all session by the given ids
      operationId: revokeAllByIds
      security:
        - JwtToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_session_v1+json:
            schema:
              $ref: '#/components/schemas/IdsRequestV1'
      responses:
        200:
          description: Sessions revoked
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user:
    post:
      tags:
        - user
      description: Create a user
      operationId: create
      security:
        - JwtToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/UserRequestV1'
      responses:
        200:
          description: User created
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
    get:
      tags:
        - user
      description: Find user information by his jwt token
      operationId: findCurrent
      security:
        - JwtToken: [ ]
      responses:
        200:
          description: User created
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user/{user_id}/role:
    post:
      tags:
        - user
      description: Add roles to a user
      operationId: addRoles
      parameters:
        - name: user_id
          in: path
          required: true
          description: User id
          schema:
            type: string
      security:
        - JwtToken: [ ]
        - AccessToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/IdsRequestV1'
      responses:
        200:
          description: Roles added
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
    delete:
      tags:
        - user
      description: Remove roles of a user
      operationId: removeRoles
      parameters:
        - name: user_id
          in: path
          required: true
          description: User id
          schema:
            type: string
      security:
        - JwtToken: [ ]
        - AccessToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/IdsRequestV1'
      responses:
        200:
          description: Roles removed
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user/{user_id}:
    get:
      tags:
        - user
      description: Find user information by his id
      operationId: findById
      parameters:
        - name: user_id
          in: path
          required: true
          description: User id
          schema:
            type: string
      security:
        - JwtToken: [ ]
        - AccessToken: [ ]
      responses:
        200:
          description: User found
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user/password:
    patch:
      tags:
        - user
      description: Update user password
      operationId: updatePassword
      security:
        - JwtToken: [ ]
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/PasswordRequestV1'
      responses:
        200:
          description: Roles added
          content:
            application/vnd.kaiqkt_auth_api_user_v1+json:
              schema:
                $ref: '#/components/schemas/UserResponseV1'
        400:
          description: Invalid payload
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
...
  /user/verify-email:
    get:
      tags:
        - user
      description: Verify user email
      operationId: verifyEmail
      parameters:
        - name: code
          in: query
          required: true
          description: Verification code
          schema:
            type: string
      responses:
        200:
          description: Email verified
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Email verified</body></html>"
        404:
          description: Email verification code not found
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Email verification failed</body></html>"
  /user/reset-password:
    post:
      tags:
        - user
      description: Reset user password request
      operationId: resetPasswordRequest
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/SendResetPasswordRequestV1'
      responses:
        200:
          description: Reset password request sent successfully
        404:
          description: User not found
          content:
            application/vnd.kaiqkt_auth_api_error_v1+json:
              schema:
                $ref: '#/components/schemas/ErrorV1'
  /user/{code}/reset-password:
    post:
      tags:
        - user
      description: Reset user password updated
      operationId: resetPassword
      parameters:
        - name: code
          in: path
          required: true
          description: Code for update password by a reset password request
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/vnd.kaiqkt_auth_api_user_v1+json:
            schema:
              $ref: '#/components/schemas/ResetPasswordRequestV1'
      responses:
        200:
          description: Password update successfully
  /user/{code}/reset-password-form:
    get:
      tags:
        - user
      description: Reset password form
      operationId: resetPasswordForm
      parameters:
        - name: code
          in: path
          required: true
          description: Code for update password by a forget password request
          schema:
            type: string
      responses:
        200:
          description: Form to reset password
          content:
            text/html:
              schema:
                type: string
                example: "<html><body>Form</body></html>"
components:
  schemas:
    ResetPasswordRequestV1:
      type: object
      properties:
        new_password:
          type: string
          example: "password"
          x-field-extra-annotation: '@field:Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).{8,}$", message = "password must be at least 8 characters long and include at least one letter, one special character, and one number")'
      required:
        - new_password
    SendResetPasswordRequestV1:
      type: object
      properties:
        email:
          type: string
          example: "test@email.com"
      required:
        - email
    PasswordRequestV1:
      type: object
      properties:
        new_password:
          type: string
          example: "password"
          x-field-extra-annotation: '@field:Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).{8,}$", message = "password must be at least 8 characters long and include at least one letter, one special character, and one number")'
        old_password:
          type: string
          example: "password"
      required:
        - new_password
        - old_password
    UserResponseV1:
      type: object
      properties:
        id:
          type: string
          example: "01E5ZQZQZQZQZQZQZQZQZQZQZ"
        first_name:
          type: string
          example: "John"
        last_name:
          type: string
          example: "Pork"
        email:
          type: string
          example: "test@email.com"
        is_verified:
          type: boolean
          example: false
        roles:
          type: array
          items:
            $ref: '#/components/schemas/RoleResponseV1'
        created_at:
          type: string
          example: "2021-01-01T00:00:00Z"
        updated_at:
          type: string
          example: "2021-01-01T00:00:00Z"
      required:
        - id
        - first_name
        - is_verified
        - email
        - roles
        - created_at
    UserRequestV1:
      type: object
      properties:
        first_name:
          type: string
          example: "John"
          x-field-extra-annotation: '@field:Size(min=1, max=255, message = "first name size must be between 1 to 255 characters")'
        last_name:
          type: string
          example: "Pork"
          x-field-extra-annotation: '@field:Size(max=255, message = "last name must not exceed 255 characters")'
        email:
          type: string
          example: "test@email.com"
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotEmpty(message = "email must not be empty") @field:jakarta.validation.constraints.Email(message = "email must be a valid address")'
        password:
          type: string
          example: "password"
          x-field-extra-annotation: '@field:Pattern(regexp = "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@#$%^&+=]).{8,}$", message = "password must be at least 8 characters long and include at least one letter, one special character, and one number")'
      required:
        - first_name
        - email
        - password
    IdsRequestV1:
      type: object
      properties:
        ids:
          type: array
          items:
            type: string
          x-field-extra-annotation: '@field:jakarta.validation.constraints.NotEmpty(message = "ids must not be empty")'
      required:
        - ids
...
    AuthenticationResponseV1:
      type: object
      properties:
        access_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
        refresh_token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IkpvaG4gRG9lIiwiaWF0IjoxNTE2MjM5MDIyfQ.SflKxwRJSMeKKF2QT4fwpMeJf36POk6yJV_adQssw5c"
      required:
        - access_token
        - refresh_token
...
    RoleResponseV1:
      type: object
      properties:
        id:
          type: string
          example: "01E5ZQZQZQZQZQZQZQZQZQZQZQ"
        name:
          type: string
          example: "ADMIN"
        description:
          type: string
          example: "Role for administrator purpose only"
        created_at:
          type: string
          example: "2021-01-01T00:00:00Z"
      required:
        - id
        - name
        - created_at
...
    ErrorV1:
      type: object
      properties:
        type:
          type: string
          description: The type of error.
        message:
          type: string
          nullable: true
          description: The error message.
        errors:
          type: object
          additionalProperties:
            type: object
          nullable: false
          description: A map of additional error details.
      required:
        - type
  securitySchemes:
    JwtToken:
      type: http
      scheme: bearer
    AccessToken:
      type: apiKey
      name: Authorization
      in: header